+++ src/game_version.js
@@ -1,7 +1,7 @@
 // @ts-check
 "use strict";
 
-const game_version = "v0.7.53";
+const game_version = "v0.7.54";
 
 function get_game_version() {
     return game_version;